- Now the name of the coins appears in the balances, instead of "unknown";

+++ app/src/main/java/cy/agorise/crystalwallet/dao/CryptoCurrencyDao.java
@@ -20,7 +20,7 @@ public interface CryptoCurrencyDao {
     List<CryptoCurrency> getAll();
 
     @Query("SELECT * FROM crypto_currency WHERE id = :id")
-    CryptoCurrency getById(int id);
+    CryptoCurrency getById(long id);
 
     @Query("SELECT * FROM crypto_currency WHERE id IN (:ids)")
     List<CryptoCurrency> getByIds(List<Long> ids);
